Graphics engine refactoring
window_t and derived classes are now linked to the gfxengine_t when
constructed, and the old init() method has been removed.

The "New Player" state creates its menu in the intended fashion,
instead of being instantiated automatically as part of the state.

Removed some obsolete #if 0'ed out code.