Fixed map/radar rendering and scaling
Refactored the graphics engine scale handling a bit. blit() from an
offscreen window now considers the scale of that window, as well as the
scale of the target window, and both can be changed from the initial
scale inherited from the parent gfxengine_t.

The radar display makes use of the above to render the map offscreen
at 1 pixel per tile, regardless of display resolution, and then scaling
in to the intended 2x2 "native" pixels per tile when rendering into the
radar window.

Closes #28.