Added special "desktop resolution fullscreen" mode
Added mode VMID_DESKTOP, which uses the SDL2 SDL_WINDOW_FULLSCREEN_DESKTOP
flag to set a fullscreen mode with the current desktop resolution. This is
now the default video mode!

Also added mode VMID_FULLWINDOW, which is supposed to open a maximized
borderless window, but that doesn't seem to work, so it's not in the GUI
mode list at this point.

Also fixed some issues with texture unloading/reloading. (Now, we just
assume that all textures are lost when hiding the display, so we unload
them first...)

Closes #38.