Stage/level/whatever terminology refactor
The game now (currently) has 5 Regions, each with 10 Levels. Each region
is set near/above a different planet, and has a unique theme. The levels
in each region are intended to start in high orbit, and subsequent levels
lay closer to the planet, the final one being just above the surface.

The game logic still counts the number of times all maps have been beaten,
internally refered to as "restarts," scaling up the difficulty with each
restart.

The dashboard now displays STAGE/REGION/LEVEL when in-game.