New skill levels: Normal, Hard and Insane
Removed the old skill levels, and replaced them with Normal, Hard and
Insane. The game speed is now constant, and the difference is in shield
bonuses awarded, and damage made by player bullets.

More tweaking needed, but the harder levels now give a better feeling of
bases putting up actual resistance, and every bullet counting. Defensive
strategies are now required when dealing with Boomerang fighters, as it
takes longer to destroy bases, and the red bullets make a lot of damage.

Closes #82.