Implemented realtime dithering for the spinning planet
There are now various static and temporal dither modes, as well as
a TrueColor mode, for those who can't handle any of the dithering.

The different planet texture variants have been replaced with a
single one, based on one of the scrolling backgrounds. Brightness
and contrast controls are applied during initialization, to allow
the planet to look darker when in the distance.

Closes #101.