Added extrapolation as an alternative to "tweening"
Added a motion filtering mode that extrapolates from the current state,
based on the current and previous state, instead of interpolating between
them. This mode is now the default.

(Unfortunately, cs.[ch] doesn't have access to the velocity information of
the game logic, or we could just have used that. #91 would fix this.)

Closes #16.