Fixed music control logic
KOBO_sound will now remember which song (if any) is supposed to be
playing, and through which group, and will stop, start and restart as
needed after configuration changes and API requests. It also filters
out attempts to start the song that's already playing.

Cleaned out some unused code.

Rudimentary logic for per-area songs in place, but we're probably going
to do something a bit more dynamic than that...

Closes #127.