Migrated to the new Audiality 2 timestamping API + minor tweaks
The Audiality 2 timestamping API has been refactored and fixed, and this
commit adds support for it.

By default, the API clock will be "synchronized" with game logic time so
that the worst case margin for messages arriving to the engine is 5 ms
('tsdelay') before deadline, regardless of latency or video frame rate.

The default sample rate has been changed from 44.1 to 48 kHz, and the
default latency has been changed from 50 to 20 ms.

Added 'tsdebug' options, that prints API message arrival timing statistics
collected bf the engine. (A2_PTSMARGIN* properties.)

A2_PTSMARGINMIN is monitored, and for each frame with late messages, the
API timestamp is bumped an extra millisecond, to correct for cases where
we start with incorrect timing. (The nudge logic would take care of that
too, but it takes a while for an error of tens of ms...)

No music started when about to exit the game.

Closes #133.