Themes are now refered to by directory; not theme file
A theme is now a "bundle" directory that is expected to contain a theme
file named "main.theme". The public API load_theme() now takes a platform
independent path (no longer a system path!) to a theme directory.

This change also applies to the 'fallback' directive.

The loader, mono, and redux themes have been updated accordingly.

Closes #155.