Removed the broken SFont spacing adjustment
No longer using ExtraSpace() in gfxengine_t::loadfont(), as it wasn't
doing the right thing anyway.

Removed the 'srcscale' argument from gfxengine_t::loadfont(), as that
information should already be in gfxengine_t.

NOTE: Note sure if this does the right thing for SFonts with scaled source
images instead of realtime scaling - but then again, what's the deal with
ExtraSpace() defaulting to 1 in the first place!?

Closes #159.