Fixed SoFont spacing
SoFont::ExtraSpace() now defaults to 0 instead of the rather strange 1, in
order to avoid problems with scaled source images. With a default of 0,
spacing data is scaled along with the source image, so applications no
longer have to compensate by scaling the ExtraSpace() value.

Unfortunately, this means all SFonts have to be fixed! They're all encoded
without spacing, relying on this weird default to cover that.

Closes #160.