Updated focus effect: Scrolling 3D grid
Added gridfx3.png to the redux theme, and changed the "plasma" effect code
to render a scrolling 3D grid with per-scanline alpha modulation for a
smooth background/effect transition at the edges. The noisy border lines
have been removed.

Brightness is modulated between 50% and 100% depending on the size of the
focus effect; it gets brighter when narrower, as that's more appropriate
for menu focus, whereas 50% brightness is better as a background for
credits, instructions etc.

Closes #131.