Player is invulnerable after destroying the last core on a stage
After destroying the final core on a stage, the player now becomes
invulnerable, until starting on the next stage. A translucent shield
effect is added to the player ship while it's invulnerable.

Note that we do NOT inhibit controls or guns at this point; it's still
possible to score a few bonus points before leaving the stage.

Closes #164.