New player ship with 64 rotation frames and smooth rotation
Added kobo-mk1c.png to the redux theme.

The player ship now bypasses normal rendering, and uses a "callback"
from kobo_gfxengine_t::post_render() instead. (This should be an engine
feature...)

The player ship rotation is filtered at the rendering frame rate, for
smooth rotation in the nearest direction to the actual heading.

Shield animatio uses manage.game_time() instead of SDL_GetTicks(), so that
the effect pauses when the game is paused.