alphamod() now activates blending for non-textured primitives
point(), fillrect(), rectangle(), fillrect_fxp(), and clear() will now
automatically enable alpha blending if alphamod() != 255, in order for it
to work like it does with the sprite*() and blit() calls.

Note, however, that this cannot be combined with other blending modes! If
any blending mode other than COPY is set, the blending mode is unchanged,
and if it's not ALPHA, that means the alphamod() value will be ignored.

Closes #184.