Ingame starfield rendered beyond the planet only
Added bool argument 'distant_only' argument to KOBO_Starfield::init(),
which sets the starfield up to render only around and beyond the planet.

Fixed (?) a rounding issue in the planet lens calculation.

Ingame starfield is now rendered with 50% alpha, to work better with the
space backdrop, and to reduce brightness. (The full color range is mapped
to the z range of the starfield at all times!)

Closes #172.