Sound themes
Sound effects and music are now themable, in a similar fashion to the
graphics themes.

The main theme can be selected via the 'sfxtheme' config string. The
default main theme is "SFX>>redux".

A fallback theme "SFX>>mono" (not yet created) can be loaded before the
main theme by setting the config switch 'force_fallback_sfxtheme'.

The loader/jingle screen has its own sound theme with a single export:
S_OAJINGLE.

Sound theme are implemented all in A2S. (No theme file parser involved.)
A theme is a directory, where the loader expects to find a main theme file
"main.a2s", which is in turn expected to import any additional files, and
export programs matching the S_* names defined in sound.h.

When the "soundtools" config switch is enabled, the loader will list any
remaining undefined sounds after loading each sound theme, which means one
doesn't have to decipher sound.h, or manually test for missing sounds.

The sound symbols used in the name code have been shortened from SOUND_*
to the same S_* format used for the symbols that the themes are expected
to export.

The KOBO_sound now has 8 banks for loading and managing multiple themes,
and also tracks which bank each sound/song belongs to, so that things are
(hopefully) killed, replaced, and restarted as expected when reloading
sounds, having a theme replace sounds from the fallback theme etc.

The 'soundtools' reload feature now (unfortunately, in most cases) reloads
everything - not just the game sound effects - as it's harder to separate
things now. Easiest workaround in case of slow reloads is to comment out
expensive songs in the theme while working on the sound design. (This
should really be done with an Audiality 2 authoring tool that can connect
to the live engine, so there's not much point in putting a lot of effort
into game engine hacks.)

Closes #222.