Delayed detonation and trigger sound effect for bombs
Bombs now have an inhibit timer between the trigger and the actual
detonation. When the player is detected, the timer starts running, while a
new sound effect, S_BOMB_TRIG (continuous) plays. If the player escapes,
the trigger warning sound stops, and the bomb goes back to scanning mode.
If the timer runs out, the bomb detonates, playing the S_BOMB_DETO sound
effect.

Minor refactoring: There is now a proper API for continous sound effects,
so that, as an alternative to the fully automatic KOBO_enemy_kind::sound
feature, enemy code can use startsound() to start a continous sound effect
manually, and have that sound managed automatically in the usual fashion.
There is also a new stopsound() call that can be used to stop and detach a
playing continuous sound effect before the enemy is killed.

Closes #226.