REFACTOR: Cleaned up kobo_form_t and menus
Replaced big(), medium(), and small() with font(), which either takes a
graphics bank index (which should of course refer to a bank holding a
font), or 0 (default argument), which uses the per-widget default font.

Added title(), which is a label() with a different font. (Well, they're
actually the same font in the current theme.)

space() takes a float argument, instead of the former integer argument.
(A feature I turned out not using this time, but oh well...)

space() accepts negative values as virtual pixels, whereas positive values
are still based on the current font, or the default: B_NORMAL_FONT.

Removed the hardcoded line heights and spacings. It's now all based on
the actual fonts.

Updated all game and options menus accordingly.