Added Debug Options menu
Added a Debug Options menu under Options, and refactored the affected
features as needed;

Added OS_RELOAD_SOUNDS. (force_fallback_sfxtheme)

Replaced display_t dfps with a plain printout. (show_fps)

The Sound Design Tools GUI is built and shown if 'soundtools' is turned
on. If 'soundtools' is turned off, the GUI is hidden.

KOBO_map::mark_edges() (which is applied when building the map) replaced
with cyan lines along the top and left edges of each map layer.

NOTE: The debug menu does not cover the 'toolstheme' path option.

Closes #247.