Config parser now ignores and drops unknown keys
Instead of giving up and ignoring the rest of the config file when hitting
an unrecognized key, the config parser now prints a warning, and skips to
the next line in the file.

NOTE: This has no effect on command line switches. There's no reliable way
of skipping switches, and it doesn't seem like a desirable feature anyway.

Closes #261.