Path and directory structure refactoring
On Un*xne, the user config file is now found in "~/.koboredux", which is
automatically created if it doesn't exist when saving the config. This
directory is also added to the THEMES>> paths, which means it can contain
sound and graphics themes in sfx and gfx subdirectories.

The configuration file is now named "koboredux.cfg" on all platforms.

GFX>> and SFX>> are now defined in terms of THEMES>>, which by default
covers the system install data directory, the user data directory (where
config and score files are found), and a local "data" directory, for
running inside build trees.

Path config cleanup: There are now only three optional paths; "data",
"themes", and "scores". If specified, these are added to the respective
paths.

KOBO_SYSCONFDIR now works on Un*x again, so admins can install a default
config "/etc/koboredux/default-config" that will be used when users start
the game with no "~/.koboredux/koboredux.cfg".

Fallback logic fixed; fallback themes are now loaded automatically if the
requested themes cannot be loaded. This allows us to set the default
themes to "redux" in all builds, so people building from source don't have
to hack their config to use the proprietary themes.

Refactored the related symbols to use #cmakedefine.

Implemented filemapper_t::mkdir() on Un*x. (For creating "~/.koboredux".)

Closes #266.