Fixed the instructions
Using KOBO_myship::bolt_frame() for the "how to fire" animation, and the
bolts are now fired from the correct positions, and rendered on top of the
ship, like in the game.

B_BIGSHIP animation cut to 13 frames, to match the "mono" theme animation.
(This will probably not work with the final "redux" theme...)

Closes #270.