Bombs are slightly quicker to detonate on the Normal skill level
Tried various valus for 'bomb_delay' on the Normal and Hard skill levels,
but settled for just changing from 10 to 8 on Normal. The difference may
seem subtle, but with 10, it was too easy to just shoot the bombs before
they had a chance to detonate. With 8, it's significantly harder to shoot
the bombs without appoaching them diagonally, which is the whole point
with their limited sensor coverage.

Closes #282.