Added support for demo builds
Added CMake option DEMO_BUILD, which defaults to OFF. When enabled, it
defines KOBO_DEMO.

Added cfg-all-demo script.

Added flashing "DEMO" label on the title screen in demo builds. (This is
the only difference in the demo builds with this patch; just to verify
that KOBO_DEMO is defined as intended.)

Closes #326.