Nicer dependency install-* scripts
install-audiality2.sh now aborts if Audiality 2 development files are
already installed in the system. If the repository is already in place,
the script performs a 'git pull' on it, instead of trying (and failing)
to clone it.

install-full-data.sh now checks if the repository and symlink are present,
acts accordingly.

install-demo-data.sh and install-win32-redist.sh now abort if the
respective directories are already in place.

The scripts now have variables at the top for archive names and versions.