Proper itch.io/universal Linux packages
Travis CI and build-packages now install the latest stable SDL2 and
(minimal PNG-only) SDL2_image from source. These libs are stripped and
included in the itch.io/Linux tar.bz2 packages.

The tar.bz2 archives from CPack are flattened, and SDL2 libs, an itch.io
manifest, a .desktop file, and a wrapper script is added, so that these
archives can be unpacked and used as is, either manually, or via the
itch.io app.

Closes #361.