No .desktop and .sh in tar.bz2 packages when linking statically
The wrapper script is only needed when bundling custom libs, and the
.desktop file is only needed to launch the script in graphical desktop
environments. Also, these files look confusing in some environments, and
they don't even work everywhere, so they're a source of trouble and
confusion.

Now, when linking the executable statically, we can drop both of these
files, and just launch the executable directly, whether we're in the
itch.io app, in the console, or in a graphical desktop environment.

Closes #372.