Specifying -nofullscreen now forces windowed mode
Formerly, the VMID_DESKTOP and VMID_FULLWINDOW (pseudo) video modes had
the graphics engine ignore the fullscreen option, which the result that
specifying "-nofullscreen" on the command line, or hacking it to 0 in
the config, had no effect.

Now, the combination of either of these modes, and fullscreen == 0 results
in a switch to the "best" windowed mode, using the same algorithm as when
switching to windowed mode with Alt+Enter.

Closes #362.