Now actually using CONFIG>>
The CONFIG>> "class" (whatever I was thinking when I came up with that
term) was not actually used! This became blatantly obvious after removing
some of the defines that were hardwired into the code instead.

Now, we use CONFIG>> for both reading and writing config files, except for
the default-config file in the sysconfig directory on Un*x platforms.

Also added an optional **path argument to filemapper_t::fopen(), so it can
return the system path it actually comes up with. Used for logging.