Fixed KOBO_replay buffer allocation
Initial replay buffer size was still at 256 bytes, for testing the
reallocation logic. Changed to 131072 (~65 minutes).

KOBO_replay is now instantiated with no replay buffer. The buffer is
instead allocated to KOBO_REPLAY_BUFFER bytes on the first write(), and
can be reallocated as needed. (Not likely ever needed, but...)

Added a compact() method, which reallocates the replay data buffer memory
block to the size of the actual data. punchin() reallocates back up to
KOBO_REPLAY_BUFFER, to avoid having a bunch of reallocations early on, if
the overwritten replay was very short.

Added check for empty REPD chunks in the loader. (Not really an error as
such, but save() isn't supposed to generate those, and size 0 would break
the code on some platforms, because of realloc() behavior.)

Addde log messages about buffer allocation issues.