Controller input disabled when game or replay is in background
Added _manage::background(), which is used by the UI state machine to
inform the game core when the game is in the background, when menues and
similar are pushed. This method automatically pauses the game when
appropriate. (Replays keep running.)

Controller input is muted when background() == true, to avoid that UI menu
operation interferes with gameplay (which it shouldn't, because the game
is supposed to be paused in such situations) and replay control.

Closes #386.