Disabled radar in the credits and options states
Now using GS_TITLE with select_stage() in the appropriate places, to keep
the radar display turned off whenever we're not playing or selecting a
start level. (Which is actually the case when entering the main menu, even
with the start level selector disabled, because we're implicitly selecting
stage 1!)

Bonus: There are confused enemies flying all over the place; not just on
the normal intro screen. (Except on the main menu, because stage 1 doesn't
have any visible enemies instantiated before entry.)

Closes #413.