More robust start level selection
As game start level selection was still relying on manage.select_stage(),
it was possible to accidentally start a new campaign on the title map
(currently stage 14) due to the transition + delayed switch kicking in at
the wrong moment.

Added an extra interface to st_game, so stage selection can be handled in
the same fashion as save slot and skill level selection, meaning it's
reliably applied right when actually starting the new game.

Also fixed some missing initializations. (Shouldn't be needed. Yet...)

Closes #434.