Fixed campaign slot selection in demo builds
Demo builds were using the wrong API for game setup, and the most obvious
side effect of that was that demos would not save campaigns unless one
actually moved the cursor to properly select a slot in the selector. Just
hitting enter/fire/... would not select a slot, so there would be no save.

Closes #425.