Logging improvements, and hardwired verbose debug log file
Added support in the logger for having log levels send to multiple
targets.

Added "LOG>>" which resolves to "CONFIG>>log"; that is, next to the
"saves" directory.

All log output (all levels) is now sent to an extra log target that is
written to a file specified by KOBO_DEBUGLOGFILE; currently
"LOG>>debuglog.txt", which resolves to "CONFIG>>log/debuglog.txt".

Added the ability to log both to console (plain text and ANSI), and to
file (plain text, ANSI, or HTML), or to disable logging altogether.
(Apart from the hardwired debug log; that cannot be disabled.)

Closes #437.