Added space bar as "activate" in GUI
Added space as an alternate "select" key in gamecontrol_t::mapsrc().

This should be safe if the user configures "space" as a fire button, as
control key bindings override any of these hardwired bindings. (And all
fire buttons serve as "select" buttons when in the menus as well, so even
that would still work as expected, except that you can't enter the main menu
ingame using "space" bar, if it doubles as a fire button.)

Closes #468.