Added basic UI mouse support
Forms now support selection of widgets via coordinates, and this is used
to have the selection/highlight chase the mouse pointer whenever whenever
it is visible.

Mouse pointer is now always active, but is hidden until mouse motion
events are actually received, and hidden again if no mouse motion events
are received for a certain time, configurable via the new config option,
mouse_hidetime.

Mouse buttons now issue BTN_PRIMARY, BTN_SECONDARY, and BTN_TERTIARY
respectively, to both gamecontrol and gsm.

Middle and right mouse buttons are wired to decrease values in widgets
that support it, whereas left mouse remains an "activate" button, which is
(as always) interpreted as "increase."

When entering a new menu while the mouse cursor is visible, any widget
under the mouse cursor is automatically selected, as if there had been a
mouse move event. (Otherwise, it feels like the UI just forgets about the
mouse when switching to a different menu.)

Closes #246.