Disabled mouse control in replay/rewind
Added gamecontrol_t::mouse_mute(), which is used for disabling ingame
mouse input decoding when it is not desired.

Refactored _manage::player_ready(), so it can be used for entering the game
from replay/rewind as well as for entering a new game.

Mouse input is now disabled in replay/rewind mode, and reentering the game
is instead handled on the game state level, which currently honors only
the mouse buttons, to avoid confusing navigation weirdness.

Closes #489.