Migrate config and saves from pre 0.7.5 installs
When finding a config file older than 0.7.5 (which means 0, because there
was no 'version' key before 0.7.5), write a new config file, and then load
and resave any campaign saves.

(Note that due to the way the campaign file backup feature works, it will
actually write backups of the original saves in the new location as a
bonus feature.)

The main reason for doing this is that Windows builds used to store those
files inside the game install directory, where they are at risk of being
deleted or overwritten either manually, or by installers and the like.

Closes #458.