Fixed GUI sound effects when selecting using the mouse
Replaced kobo_form_t::prev() and next() with select(), which covers both,
and also works when selection is done with mouse or similar, rather than
buttons or keys.

ct_form_t::select() now does nothing if the widget is already selected, to
avoid spamming sound effects and other things when using the mouse.

Closes #525.