Fixed missing mouse control in certain situations
Added "Return to Intro" button in the campaign selector.

Renamed BTN_EXIT as BTN_MENU, as that's what it's almost exclusively used
for.

Added BTN_LMB, BTN_MMB, and BTN_RMB, which are used for explicitly
trapping mouse buttons in the GUI when ingame mouse control is enabled.
When 'mouse' (ingame mouse control) is disabled, the mouse buttons are
translated to BTN_SELECT (LMB) and BTN_MENU (MMB/RMB).

Most states treat mouse buttons the same whether or not ingame mouse
control is enabled, except when ingame, where the GUI ignores the mouse
buttons.

Closes #529.