Replaced "Continue?" state with delayed automatic rewind/retry
Added 'autocontinue' option, default 5, which overrides the "Continue?"
dialog with a plain delay, that can be skipped after half of the delay has
elapsed.

This should also improve the #484 situations, as there is now a plain,
non-confusing delay before entering the rewind/retry mode.

By setting 'autocontinue' to 0, the old "Continue?" dialog is brought
back, and is configurable like before.

Closes #535.
Closes #484.