Arcade style demo mode
Added arcade style demo playback on the title screen, playing randomly
selected level replays out of campaign saves stored in "DEMOS>>". (That
is, "demos" anywhere in the "DATA>>", where one finds "gfx" and "sfx".)

The ingame graphics from the demos is faded to 50% intensity when the
title "pages" are showing, or when a menu is up, and something in it is
highlighted.

Demo playback can be disabled via the 'titledemos' option, in which case
the old slow scrolling map is used instead. This also happens if no
compatible demos are found.

Various refactoring and tweaks had to be done to make all this work.

A single random campaign replay is included as "demos/demo_0.dat".

Closes #523.