Automatically redirect stdout and stderr to console on Win32
Win32 builds now attempt to connect to an existing console window first
thing, so that copyright, library versions etc can be seen, and so that
stdout and stderr automatically work when starting the game from console.

The game will still attempt to open a new console window after loading
the configuration file, if debug is enabled by that configuration file.

This change should have no effect on other platforms.

Closes #555.